==> notebooks/4-experiment_04_merging_data.ipynb <==
import os
import mlflow
import mlflow.sklearn
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix, ConfusionMatrixDisplay
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

project_root = os.path.abspath(os.path.join(os.getcwd(), ".."))
processed_dir = os.path.join(project_root, "notebooks", "data")
df_test = pd.read_csv(os.path.join(processed_dir, "test_preprocessed.csv"))
df_train = pd.read_csv(os.path.join(processed_dir, "train_preprocessed.csv"))

full_df = pd.concat([df_train, df_test], ignore_index=True)
full_df.head()

full_df.shape

full_df["Churn"].value_counts()

from sklearn.utils import shuffle
full_df = shuffle(full_df, random_state=42).reset_index(drop=True)
full_df.head()

X = full_df.drop(columns=["Churn"])
y = full_df["Churn"]
X_train_full, X_test, y_train_full, y_test = train_test_split(
    X, y, 
    test_size=0.2, 
    random_state=42, 
    stratify=y
)
X_train, X_valid, y_train, y_valid = train_test_split(
    X_train_full, y_train_full, 
    test_size=0.25, 
    random_state=42, 
    stratify=y_train_full
)
scaler = StandardScaler()
scaler.fit(X_train)

X_train_scaled = scaler.transform(X_train)
X_valid_scaled = scaler.transform(X_valid)
X_test_scaled = scaler.transform(X_test)

X_train_scaled = pd.DataFrame(X_train_scaled, columns=X_train.columns)
X_valid_scaled = pd.DataFrame(X_valid_scaled, columns=X_valid.columns)
X_test_scaled = pd.DataFrame(X_test_scaled, columns=X_test.columns)

X_train_scaled.to_csv(os.path.join(project_root , "notebooks","data", "X_train_scaled_merged.csv"), index=False)
y_train.to_csv(os.path.join(project_root , "notebooks","data", "y_train_merged.csv"), index=False)

X_valid_scaled.to_csv(os.path.join(project_root , "notebooks","data", "X_valid_scaled_merged.csv"), index=False)
y_valid.to_csv(os.path.join(project_root , "notebooks","data", "y_valid_merged.csv"), index=False)

X_test_scaled.to_csv(os.path.join(project_root , "notebooks","data", "X_test_scaled_merged.csv"), index=False)
y_test.to_csv(os.path.join(project_root , "notebooks","data", "y_test_merged.csv"), index=False)



print("\nSplitting and scaling complete!")
print(f"Final Training set shape: {X_train_scaled.shape}")
print(f"Validation set shape: {X_valid_scaled.shape}")
print(f"Test set shape: {X_test_scaled.shape}")
